refactor(wallet-sdk): move session-keys into domain/sdk
